third_party/WebKit/LayoutTests/platform/win/http/tests/security/subresourceIntegrity/subresource-integrity-invalid-integrity-expected.txt

CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256;B0_62fJSJFrdjEFR9ba04m_D-LHQ-zG6PGcaR0Trpxg='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256:B0_62fJSJFrdjEFR9ba04m_D-LHQ-zG6PGcaR0Trpxg='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
CONSOLE ERROR: Error parsing 'integrity' attribute ('sha256-(B0_62fJSJFrdjEFR9ba04m_D-LHQ-zG6PGcaR0Trpxg=)'). The digest must be a valid, base64-encoded value.
This is a testharness.js-based test.
PASS SRI invalid integrity attributes 
Harness: the test ran to completion.
